# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FoodTinder API",
        "description": "API do logowania, swipe’ów i rekomendacji dań. Po prostu Tinder dla jedzenia!",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        }
    ],
    "paths": {
        "api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logowanie",
                "operationId": "0aabac24e722c84ecb25cbe844514866",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/LoginRequest"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthTokenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Wylogowanie",
                "operationId": "51c1fe6ecf4e5fb1424d17d9606d99a2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/verify-email/{id}/{hash}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Potwierdzenie adresu email",
                "operationId": "36da40610f39b3888902c5a1b2010396",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Email already verified"
                    },
                    "200": {
                        "description": "Email successfully verified"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthUser": {
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "name": {
                        "type": "string",
                        "example": "Jan Kowalski"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jan@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-17T19:52:15.000000Z"
                    },
                    "status": {
                        "type": "string",
                        "example": "verification-link-sent"
                    }
                },
                "type": "object"
            },
            "AuthTokenResponse": {
                "required": [
                    "token",
                    "user"
                ],
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "1|abcdefgh..."
                    },
                    "user": {
                        "$ref": "#/components/schemas/AuthUser"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jan@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "secret123"
                    }
                },
                "type": "object"
            },
...
            "MessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Successfully logged out"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "schema": "AuthSchemas",
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}